Build scrapix into a cjs module
