+++ src/server.ts
@@ -2,9 +2,9 @@ import * as dotenv from 'dotenv'
 dotenv.config()
 
 import express from 'express'
-import TaskQueue from './taskQueue.js'
+import { TaskQueue } from './taskQueue.js'
 import { Sender } from './sender.js'
-import Crawler from './crawler.js'
+import { Crawler } from './crawler.js'
 
 const port = process.env.PORT || 8080
